Set the type of the images from double to short. Images are usually using short for pixel value and memory can be saved by using it instead of double type.

+++ Utilities/btkNrrdToNifti.cxx
@@ -51,7 +51,7 @@ knowledge of the CeCILL-B license and that you accept its terms.
 #include "vnl/vnl_inverse.h"
 
 
-typedef double PixelType;
+typedef short PixelType;
 const unsigned int InDimension  = 3;
 const unsigned int OutDimension = 4;
